Organizing integrators using local _utils.py
