+++ tests/integrators/test_solver_options.py
@@ -20,10 +20,10 @@ from jaxfun.galerkin.functionspace import FunctionSpace
 from jaxfun.galerkin.Legendre import Legendre
 from jaxfun.galerkin.tensorproductspace import TensorProduct
 from jaxfun.integrators import ARK4_3_6L2SA, BackwardEuler, IMEXRungeKutta
-from jaxfun.integrators.base import (
-    _accepted_solve_options,
-    _validate_solver_options,
+from jaxfun.integrators._utils import (
+    accepted_solve_options,
     known_solve_options,
+    validate_solver_options,
 )
 from jaxfun.la import IdentityMatrix, Matrix
 from jaxfun.la.tpmatrix import TPMatrices
@@ -80,28 +80,28 @@ def _diffusion_2d(**params):
 
 def test_only_declared_options_are_offered_to_an_operator() -> None:
     """Each operator sees the keyword-only options its own `solve` names."""
-    assert _accepted_solve_options(TPMatrices) == {
+    assert accepted_solve_options(TPMatrices) == {
         "method",
         "kron_method",
         "auto_threshold",
     }
     # A dense solve has nothing to choose, so it is offered nothing -- passing
     # `auto_threshold` to an integrator with a dense mass matrix is not an error.
-    assert _accepted_solve_options(Matrix) == frozenset()
-    assert _accepted_solve_options(IdentityMatrix) == frozenset()
+    assert accepted_solve_options(Matrix) == frozenset()
+    assert accepted_solve_options(IdentityMatrix) == frozenset()
     assert known_solve_options() >= {"method", "kron_method", "auto_threshold"}
 
 
 def test_options_are_normalized_to_sorted_pairs() -> None:
     """Options are stored order-independently so equal configs compare equal."""
-    assert _validate_solver_options(None) == ()
-    assert _validate_solver_options({}) == ()
-    assert _validate_solver_options({"auto_threshold": 400, "method": "lu"}) == (
+    assert validate_solver_options(None) == ()
+    assert validate_solver_options({}) == ()
+    assert validate_solver_options({"auto_threshold": 400, "method": "lu"}) == (
         ("auto_threshold", 400),
         ("method", "lu"),
     )
-    assert _validate_solver_options({"method": "lu", "auto_threshold": 400}) == (
-        _validate_solver_options({"auto_threshold": 400, "method": "lu"})
+    assert validate_solver_options({"method": "lu", "auto_threshold": 400}) == (
+        validate_solver_options({"auto_threshold": 400, "method": "lu"})
     )
